api/hotel-guest/hotel/requirements.txt
Flask==3.0.3
Flask_Cors==4.0.0
psycopg==3.1.18
psycopg_binary==3.1.18
python-dotenv==1.0.1